requirements-top-level.txt
advanced_alchemy==0.34.0
asyncpg==0.28.0
geojson==3.1.0
litestar==2.21.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-httpx==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic-settings==2.13.1
redis==5.0.1
structlog==24.1.0
uvicorn==0.34.0
